Moved Permgen Paramter Creation into Launcher Classes
I mostly used the code in getActualJavaVersion and isJava8 to create the
launcher factory, which takes a java path and figures out the rest from
there. After that, I have a launcher and I can tell that java launcher
to just create the right permgeh parameter and don't need to worry about
this anymore.